changes for lesson25 part1 navigation from directory
