+++ src/components/directory-item/directory-item.component.jsx
@@ -1,12 +1,15 @@
 // import './category-item.styles.scss';
 //import './directory-item.styles.scss';
+import { useNavigate } from "react-router-dom";
 import { BackgroundImage,Body,DirectoryItemContainer } from "./directory-item.styles";
 
 //const CategoryItem =({category}) =>{
 const DirectoryItem =({category}) =>{
-    const {title,imageUrl} = category;
+    const {title,imageUrl,route} = category;
+    const navigate = useNavigate();
+    const onNavigateHandler =() =>navigate(route);
     return(
-        <DirectoryItemContainer>   
+        <DirectoryItemContainer onClick={onNavigateHandler}>   
             <BackgroundImage
                  imageUrl={imageUrl}>
             </BackgroundImage>
