+++ src/contexts/categories.context.jsx
@@ -16,7 +16,7 @@ export const CategoriesProvider =({children})=>{
      useEffect(()=>{
          const getCategoriesMap = async() => {
              const categoryMap = await getCategoriesAndDocuments();
-             console.log(categoryMap);  
+            // console.log(categoryMap);  
              setCategoriesMap(categoryMap);          
          }
          getCategoriesMap();
